Deduplicate `ConvertToOpenApiType`
